WIP: Migration from scriptProcessor to AudioWorkletProcessor
BaseAudioContext:createScriptProcessor is deprecated for a long time now
see https://developer.mozilla.org/en-US/docs/Web/API/BaseAudioContext/createScriptProcessor
So it became mandatory to migrate to the new AudioWorklet API.

Given that their functioning is radically different, the choice was made
to completely migrate to this new interface, leaving support for old versions
on another branch.

This is a work in progress.